+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   by `Lambda()` function.
 - Adds `to_string`, `to_number`, `to_time`, and `to_date` functions
 - Pretty print Fauna objects
+- Adds `@bytes` constructor on query api
 
 ## 2.0.2 (March 28, 2018)
 - Bug fix: functions with optional scope parameter were failing when scope was